Refactor updatemodelcache pacakge
