+++ plugin/src/main/java/org/opensearch/ml/action/models/UpdateModelTransportAction.java
@@ -49,10 +49,10 @@ import org.opensearch.ml.common.model.MLModelState;
 import org.opensearch.ml.common.transport.model.MLUpdateModelAction;
 import org.opensearch.ml.common.transport.model.MLUpdateModelInput;
 import org.opensearch.ml.common.transport.model.MLUpdateModelRequest;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheAction;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodeResponse;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodesRequest;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodesResponse;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheAction;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodeResponse;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodesRequest;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodesResponse;
 import org.opensearch.ml.engine.MLEngine;
 import org.opensearch.ml.helper.ConnectorAccessControlHelper;
 import org.opensearch.ml.helper.ModelAccessControlHelper;
