+++ plugin/src/main/java/org/opensearch/ml/action/update_cache/UpdateModelCacheTransportAction.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.opensearch.ml.action.updatemodelcache;
+package org.opensearch.ml.action.update_cache;
 
 import java.io.IOException;
 import java.util.HashMap;
@@ -20,11 +20,11 @@ import org.opensearch.core.action.ActionListener;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.xcontent.NamedXContentRegistry;
 import org.opensearch.ml.cluster.DiscoveryNodeHelper;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheAction;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodeRequest;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodeResponse;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodesRequest;
-import org.opensearch.ml.common.transport.updatemodelcache.MLUpdateModelCacheNodesResponse;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheAction;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodeRequest;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodeResponse;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodesRequest;
+import org.opensearch.ml.common.transport.update_cache.MLUpdateModelCacheNodesResponse;
 import org.opensearch.ml.helper.ModelAccessControlHelper;
 import org.opensearch.ml.model.MLModelManager;
 import org.opensearch.ml.stats.MLStats;
